requirements.txt
streamlit
numpy
Pillow
tensorflow
matplotlib
joblib
torch
fastai
fastcore
